+++ client/config_test.go
@@ -5,18 +5,16 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/tests"
-
-	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/tests"
 )
 
 // For https://github.com/cosmos/cosmos-sdk/issues/3899
 func Test_runConfigCmdTwiceWithShorterNodeValue(t *testing.T) {
 	// Prepare environment
-	t.Parallel()
-
 	configHome, cleanup := tests.NewTestCaseDir(t)
 	t.Cleanup(cleanup)
 
@@ -42,8 +40,6 @@ func Test_runConfigCmdTwiceWithShorterNodeValue(t *testing.T) {
 
 func TestConfigCmd_OfflineFlag(t *testing.T) {
 	// Prepare environment
-	t.Parallel()
-
 	configHome, cleanup := tests.NewTestCaseDir(t)
 	t.Cleanup(cleanup)
 
@@ -52,11 +48,14 @@ func TestConfigCmd_OfflineFlag(t *testing.T) {
 
 	// Init command config
 	cmd := ConfigCmd(configHome)
+	_, out, _ := tests.ApplyMockIO(cmd)
 	assert.NotNil(t, cmd)
 
 	viper.Set(flagGet, true)
 	err := cmd.RunE(cmd, []string{"offline"})
 	assert.Nil(t, err)
+	assert.Contains(t, out.String(), "false")
+	out.Reset()
 
 	viper.Set(flagGet, false)
 	err = cmd.RunE(cmd, []string{"offline", "true"})
@@ -65,4 +64,5 @@ func TestConfigCmd_OfflineFlag(t *testing.T) {
 	viper.Set(flagGet, true)
 	err = cmd.RunE(cmd, []string{"offline"})
 	assert.Nil(t, err)
+	assert.Contains(t, out.String(), "true")
 }